+++ README.md
@@ -1,1 +1,10 @@
-link https://cdn-shop.adafruit.com/datasheets/SSD1306.pdf
+# Adafruit_SSD1306_128X32
+
+A Particle library for Adafruit_SSD1306_128X32
+
+## Documentation
+
+[SSD1306](https://cdn-shop.adafruit.com/datasheets/SSD1306.pdf)
+
+## LICENSE
+Copyright 2018 Lucian
