+++ library.properties
@@ -1,0 +1,5 @@
+name=SSD1306_128x32
+version=1.1.2
+author=Lucian<lucian@particle.io>
+url=https://github.com/particle-iot/adafruit_ssd1306
+dependencies.SSD1306_128x32=1.1.2
