+++ src/SSD1306_128x32.h
@@ -25,6 +25,7 @@ All text above, and the splash screen must be included in any redistribution
 #define WHITE 1
 
 #define SSD1306_I2C_ADDRESS   0x3C	// 011110+SA0+RW - 0x3C or 0x3D
+//#define SSD1306_I2C_ADDRESS   0x3D	// 011110+SA0+RW - 0x3C or 0x3D
 // Address for 128x32 is 0x3C
 // Address for 128x64 is 0x3D (default) or 0x3C (if SA0 is grounded)
 
@@ -40,8 +41,8 @@ All text above, and the splash screen must be included in any redistribution
     SSD1306_128_32  128x32 pixel display
 
     -----------------------------------------------------------------------*/
-#define SSD1306_128_64
-//#define SSD1306_128_32
+//#define SSD1306_128_64
+#define SSD1306_128_32
 /*=========================================================================*/
 
 #if defined SSD1306_128_64 && defined SSD1306_128_32
@@ -106,11 +107,11 @@ All text above, and the splash screen must be included in any redistribution
 #define SSD1306_VERTICAL_AND_RIGHT_HORIZONTAL_SCROLL 0x29
 #define SSD1306_VERTICAL_AND_LEFT_HORIZONTAL_SCROLL 0x2A
 
-class Adafruit_SSD1306 : public Adafruit_GFX {
+class SSD1306_128x32 : public Adafruit_GFX {
  public:
-  Adafruit_SSD1306(int8_t SID, int8_t SCLK, int8_t DC, int8_t RST, int8_t CS);
-  Adafruit_SSD1306(int8_t DC, int8_t RST, int8_t CS);
-  Adafruit_SSD1306(int8_t RST);
+  SSD1306_128x32(int8_t SID, int8_t SCLK, int8_t DC, int8_t RST, int8_t CS);
+  SSD1306_128x32(int8_t DC, int8_t RST, int8_t CS);
+  SSD1306_128x32(int8_t RST);
 
   void begin(uint8_t switchvcc = SSD1306_SWITCHCAPVCC, uint8_t i2caddr = SSD1306_I2C_ADDRESS);
   void ssd1306_command(uint8_t c);
